+++ utils/chainConfigs.js
@@ -1,3 +1,43 @@
+// coctohug maps to chainname on API
+const chainNameMap = {
+  default: 'chia',
+
+  cactus: 'cactus',
+  covid: 'covid',
+  cryptodoge: 'cryptodoge',
+  ethgreen: 'ethgreen',
+  flora: 'flora',
+  greendoge: 'greendoge',
+  lucky: 'lucky',
+  pipscoin: 'pipscoin',
+  shibgreen: 'shibgreen',
+  silicoin: 'silicoin',
+
+  skynet: 'skynet',
+  stor: 'stor',
+  staicoin: 'stai',
+  tranzact: 'tranzact',
+  venidium: 'venidium',
+  btcgreen: 'btcgreen',
+  hddcoin: 'hddcoin',
+  maize: 'maize',
+  flax: 'flax',
+  aedge: 'aedge',
+
+  apple: 'apple',
+  wheat: 'wheat',
+  dogechia: 'dogechia',
+  tad: 'tad',
+  taco: 'taco',
+  socks: 'socks',
+  mogua: 'mogua',
+  mint: 'mint',
+  salvia: 'salvia',
+  chia: 'chia',
+
+  nchain: 'nchain',
+};
+
 const getExp = (blockchain) => {
   return `https://alltheblocks.net/${blockchain}/address/`
 };
@@ -6,41 +46,37 @@ const getPeers = (blockchain) => {
   return `https://api.alltheblocks.net/${blockchain}/peer/recent?amount=100`;
 };
 
-const chainConfigs = {
-  default: { exp: getExp('chia'), peers: getPeers('chia') },
-  chia: { exp: getExp('chia'), peers: getPeers('chia') },
-
-  cactus: { exp: getExp('cactus'), peers: getPeers('cactus') },
-  covid: { exp: getExp('covid'), peers: getPeers('covid') },
-  cryptodoge: { exp: getExp('cryptodoge'), peers: getPeers('cryptodoge') },
-  ethgreen: { exp: getExp('ethgreen'), peers: getPeers('ethgreen') },
-  flora: { exp: getExp('flora'), peers: getPeers('flora') },
-  greendoge: { exp: getExp('greendoge'), peers: getPeers('greendoge') },
-  lucky: { exp: getExp('lucky'), peers: getPeers('lucky') },
-  pipscoin: { exp: getExp('pipscoin'), peers: getPeers('pipscoin') },
-  shibgreen: { exp: getExp('shibgreen'), peers: getPeers('shibgreen') },
-  silicoin: { exp: getExp('silicoin'), peers: getPeers('silicoin') },
-
-  skynet: { exp: getExp('skynet'), peers: getPeers('skynet') },
-  staicoin: { exp: getExp('stai'), peers: getPeers('stai') },
-  stor: { exp: getExp('stor'), peers: getPeers('stor') },
-  tranzact: { exp: getExp('tranzact'), peers: getPeers('tranzact') },
-  venidium: { exp: getExp('venidium'), peers: getPeers('venidium') },
-  btcgreen: { exp: getExp('btcgreen'), peers: getPeers('btcgreen') },
-  hddcoin: { exp: getExp('hddcoin'), peers: getPeers('hddcoin') },
-  maize: { exp: getExp('maize'), peers: getPeers('maize') },
-  flax: { exp: getExp('flax'), peers: getPeers('flax') },
-  aedge: { exp: getExp('aedge'), peers: getPeers('aedge') },
-
-  apple: { exp: getExp('apple'), peers: getPeers('apple') },
-  wheat: { exp: getExp('wheat'), peers: getPeers('wheat') },
-  dogechia: { exp: getExp('dogechia'), peers: getPeers('dogechia') },
-  tad: { exp: getExp('tad'), peers: getPeers('tad') },
-  taco: { exp: getExp('taco'), peers: getPeers('taco') },
-  socks: { exp: getExp('socks'), peers: getPeers('socks') },
-  mogua: { exp: getExp('mogua'), peers: getPeers('mogua') },
-  mint: { exp: getExp('mint'), peers: getPeers('mint') },
-  salvia: { exp: getExp('salvia'), peers: getPeers('salvia') },
+const getCoinRecordsUrl = (blockchain, address, pageNumber, pageSize) => {
+  const apiName = chainNameMap[blockchain];
+  return `https://api.alltheblocks.net/${apiName}/coin/address/${address}?pageNumber=${pageNumber}&pageSize=${pageSize}`;
 };
 
-module.exports = chainConfigs;
+const getBalanceUrl = (blockchain, address) => {
+  const apiName = chainNameMap[blockchain];
+  return `https://api.alltheblocks.net/${apiName}/address/name/${address}`;
+};
+
+const getMojoDivider = (blockchain) => {
+  if (['shibgreen'].includes(blockchain)) return 1000;
+  else if (['cryptodoge'].includes(blockchain)) return 1000000;
+  else if (['ethgreen', 'staicoin'].includes(blockchain)) return 1000000000;
+
+  return 1000000000000;
+};
+
+const chainConfigs = {};
+Object.keys(chainNameMap).forEach(cname => {
+  const apiName = chainNameMap[cname];
+  chainConfigs[cname] = {
+    exp: getExp(apiName),
+    peers: getPeers(apiName),
+    mojoDivider: getMojoDivider(cname)
+  };
+});
+
+module.exports = {
+  chainNameMap,
+  chainConfigs,
+  getCoinRecordsUrl,
+  getBalanceUrl,
+};