+++ utils/statusJobs.js
@@ -11,7 +11,9 @@ const {
 } = require('../jobs');
 
 const {
-  updateColdwalletCoins
+  updateHourlyColdwalletCoins,
+  updateDailyColdwalletCoins,
+  updateWeeklyColdwalletCoins,
 } = require('../controllerJobs');
 
 const everyMinute = '0 */1 * * * *';
@@ -19,6 +21,8 @@ const every5Minute = '0 */5 * * * *';
 const every30Minute = '0 */30 * * * *';
 const every1Hour = '0 0 */1 * * *';
 const every4Hour = '0 0 */4 * * *';
+const everyMidnight = '0 40 2 * * *';
+const everyMondayMidnight = '0 40 3 * * 1';
 
 const isWebController = isWebControllerMode();
 const isNotHarvester = !isHarvesterMode();
@@ -66,7 +70,7 @@ const thirtyMinuteJob = new CronJob(every30Minute, async () => {
 const oneHourJob = new CronJob(every1Hour, async () => {
   logger.info('oneHourJob start');
   if (isWebController) {
-    await updateColdwalletCoins();
+    await updateHourlyColdwalletCoins();
   }
   logger.info('oneHourJob end');
 }, null, true, 'America/Los_Angeles');
@@ -79,17 +83,31 @@ const fourHourJob = new CronJob(every4Hour, async () => {
   logger.info('fourHourJob end');
 }, null, true, 'America/Los_Angeles');
 
+const oneDayJob = new CronJob(everyMidnight, async () => {
+  logger.info('oneDayJob start');
+  if (isWebController) {
+    await updateDailyColdwalletCoins();
+  }
+  logger.info('oneDayJob end');
+}, null, true, 'America/Los_Angeles');
+
+const oneWeekJob = new CronJob(everyMondayMidnight, async () => {
+  logger.info('oneWeekJob start');
+  if (isWebController) {
+    await updateWeeklyColdwalletCoins();
+  }
+  logger.info('oneWeekJob end');
+}, null, true, 'America/Los_Angeles');
+
 const startAllJobs = async () => {
   logger.info('all jobs start');
-  // start jobs in hand mode only
-  if (!isWebController) {
-    oneMinuteJob.start();
-    fiveMinuteJob.start();
-    thirtyMinuteJob.start();
-    fourHourJob.start();
-  } else {
-    oneHourJob.start();
-  }
+  oneMinuteJob.start();
+  fiveMinuteJob.start();
+  thirtyMinuteJob.start();
+  oneHourJob.start();
+  fourHourJob.start();
+  oneDayJob.start();
+  oneWeekJob.start();
   logger.info('all jobs end');
 };