+++ changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 
-- auto add peers when it is less than 3
+- Account balance of Wallet
+- weekly report with coldwallet
+- daily report with coldwallet
+- auto add 30 peers when it is less than 3
 - fast node sync doc update
 
 ## [1.0.1] - 2021-12-12